Move Swing panel sample to dedicated package
