--- samples/ide-plugin/src/main/kotlin/org/jetbrains/jewel/samples/ideplugin/SwingComparison.kt
@@ -1,1 +1,0 @@
-package org.jetbrains.jewel.samples.ideplugin
